+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.22
 
 -   Fix: [objectp should maintain typed object #168](https://github.com/jlchmura/lpc-language-server/issues/168)
+-   Fix: [Incorrect return type for closure set to a function #170](https://github.com/jlchmura/lpc-language-server/issues/170)
 -   Fix: [arrayp sefun is not changing the flow type to array #171](https://github.com/jlchmura/lpc-language-server/issues/171)
 -   Added support for LPCDoc intersection types